Phases/Phase3/backend/pyproject.toml
[project]
name = "todo-phase3-backend"
version = "3.0.0"
description = "AI-Powered Todo Chatbot Backend - FastAPI + OpenAI Agents + MCP"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "pydantic>=2.12.5",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "openai>=1.60.0",
    "openai-agents>=0.0.7",
    "mcp>=1.2.0",
    "openai-chatkit>=1.5.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
